requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==2.3.0
Django==5.2.1
feedparser==6.0.11
google-ai-generativelanguage==0.6.4
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
